Fixes #31 and #37 where icons were not properly aligned in images
The 2x version of the button sprite had different proportions from the 1x version of the sprite, so the background-size property was incorrect and icons were not correctly positioned on some screens.  Updating to the latest Leaflet.Draw didn't fix the issue and in fact made it worse, so I just modified the sprites to match so the only difference is the 2x scale.